+++ packages/query_kit_flutter/lib/query_kit_flutter.dart
@@ -20,6 +20,7 @@ library;
 export 'package:query_kit/query_kit.dart';
 
 export 'src/mutation_state_controller.dart';
+export 'src/online_status.dart';
 export 'src/queries_builder.dart';
 export 'src/queries_controller.dart';
 export 'src/query_builder.dart';